requirement.txt
asgiref==3.5.2
cffi==1.15.1
colorama==0.4.5
cryptography==37.0.4
Django==4.0.6
Pillow==9.2.0
pycparser==2.21
python-dotenv==0.20.0
qrcode==7.3.1
sqlparse==0.4.2
tzdata==2022.1
